ECO-130: extract component TimeStamp.
